Shared address processor in cart and checkout
