+++ assets/js/base/context/cart-checkout/checkout/processor/utils.js
@@ -1,10 +1,4 @@
 /**
- * External dependencies
- */
-import isShallowEqual from '@wordpress/is-shallow-equal';
-import { pluckAddress } from '@woocommerce/base-utils';
-
-/**
  * @typedef {import('@woocommerce/type-defs/payments').PaymentDataItem} PaymentDataItem
  */
 
@@ -33,22 +27,4 @@ export const preparePaymentData = (
 		value: shouldSave,
 	} );
 	return apiData;
-};
-
-/**
- * Does a shallow compare of important address data to determine if the cart needs updating.
- *
- * @param {Object} previousAddress An object containing all previous address information
- * @param {Object} address An object containing all address information
- *
- * @return {boolean} True if the store needs updating due to changed data.
- */
-export const shouldUpdateAddressStore = ( previousAddress, address ) => {
-	if ( ! address.country ) {
-		return false;
-	}
-	return ! isShallowEqual(
-		pluckAddress( previousAddress ),
-		pluckAddress( address )
-	);
 };
