+++ assets/js/base/context/cart-checkout/index.js
@@ -3,4 +3,5 @@ export * from './shipping';
 export * from './billing';
 export * from './checkout';
 export * from './cart';
+export * from './address-processor';
 export { useCheckoutContext } from './checkout-state';
